+++ lib/src/painters/player_wave_painter.dart
@@ -1,70 +1,47 @@
 import 'package:flutter/material.dart';
 
-import '../base/utils.dart';
+import '../../audio_waveforms.dart';
 
 class PlayerWavePainter extends CustomPainter {
   final List<double> waveformData;
-  final bool showTop;
-  final bool showBottom;
   final double animValue;
-  final double scaleFactor;
-  final Color waveColor;
-  final StrokeCap waveCap;
-  final double waveThickness;
-  final Shader? fixedWaveGradient;
-  final Shader? liveWaveGradient;
-  final double spacing;
   final Offset totalBackDistance;
   final Offset dragOffset;
   final double audioProgress;
-  final Color liveWaveColor;
   final VoidCallback pushBack;
   final bool callPushback;
   final double emptySpace;
   final double scrollScale;
-  final bool showSeekLine;
-  final double seekLineThickness;
-  final Color seekLineColor;
   final WaveformType waveformType;
+
+  final PlayerWaveStyle playerWaveStyle;
 
   PlayerWavePainter({
     required this.waveformData,
-    required this.showTop,
-    required this.showBottom,
     required this.animValue,
-    required this.scaleFactor,
-    required this.waveColor,
-    required this.waveCap,
-    required this.waveThickness,
     required this.dragOffset,
     required this.totalBackDistance,
-    required this.spacing,
     required this.audioProgress,
-    required this.liveWaveColor,
     required this.pushBack,
     required this.callPushback,
     required this.scrollScale,
-    required this.seekLineThickness,
-    required this.seekLineColor,
-    required this.showSeekLine,
     required this.waveformType,
     required this.cachedAudioProgress,
-    this.liveWaveGradient,
-    this.fixedWaveGradient,
+    required this.playerWaveStyle,
   })  : fixedWavePaint = Paint()
-          ..color = waveColor
-          ..strokeWidth = waveThickness
-          ..strokeCap = waveCap
-          ..shader = fixedWaveGradient,
+          ..color = playerWaveStyle.fixedWaveColor
+          ..strokeWidth = playerWaveStyle.waveThickness
+          ..strokeCap = playerWaveStyle.waveCap
+          ..shader = playerWaveStyle.fixedWaveGradient,
         liveWavePaint = Paint()
-          ..color = liveWaveColor
-          ..strokeWidth = waveThickness
-          ..strokeCap = waveCap
-          ..shader = liveWaveGradient,
-        emptySpace = spacing,
+          ..color = playerWaveStyle.liveWaveColor
+          ..strokeWidth = playerWaveStyle.waveThickness
+          ..strokeCap = playerWaveStyle.waveCap
+          ..shader = playerWaveStyle.fixedWaveGradient,
+        emptySpace = playerWaveStyle.spacing,
         middleLinePaint = Paint()
-          ..color = seekLineColor
-          ..strokeWidth = seekLineThickness;
+          ..color = playerWaveStyle.seekLineColor
+          ..strokeWidth = playerWaveStyle.seekLineThickness;
 
   Paint fixedWavePaint;
   Paint liveWavePaint;
@@ -74,7 +51,9 @@ class PlayerWavePainter extends CustomPainter {
   @override
   void paint(Canvas canvas, Size size) {
     _drawWave(size, canvas);
-    if (showSeekLine && waveformType.isLong) _drawMiddleLine(size, canvas);
+    if (playerWaveStyle.showSeekLine && waveformType.isLong) {
+      _drawMiddleLine(size, canvas);
+    }
   }
 
   @override
@@ -85,8 +64,8 @@ class PlayerWavePainter extends CustomPainter {
       Offset(size.width / 2, 0),
       Offset(size.width / 2, size.height),
       fixedWavePaint
-        ..color = seekLineColor
-        ..strokeWidth = seekLineThickness,
+        ..color = playerWaveStyle.seekLineColor
+        ..strokeWidth = playerWaveStyle.seekLineThickness,
     );
   }
 
@@ -99,15 +78,17 @@ class PlayerWavePainter extends CustomPainter {
     }
     for (int i = 0; i < length; i++) {
       final currentDragPointer = dragOffset.dx - totalBackDistance.dx;
-      final waveWidth = i * spacing;
+      final waveWidth = i * playerWaveStyle.spacing;
       final dx = waveWidth +
           currentDragPointer +
           emptySpace +
           (waveformType.isFitWidth ? 0 : halfWidth);
-      final waveHeight =
-          (waveformData[i] * animValue) * scaleFactor * scrollScale;
-      final bottomDy = halfHeight + (showBottom ? waveHeight : 0);
-      final topDy = halfHeight + (showTop ? -waveHeight : 0);
+      final waveHeight = (waveformData[i] * animValue) *
+          playerWaveStyle.scaleFactor *
+          scrollScale;
+      final bottomDy =
+          halfHeight + (playerWaveStyle.showBottom ? waveHeight : 0);
+      final topDy = halfHeight + (playerWaveStyle.showTop ? -waveHeight : 0);
 
       // Only draw waves which are in visible viewport.
       if (dx > 0 && dx < halfWidth * 2) {